pyproject.toml
[project]
name = "lang"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain-core>=0.3.63",
    "langchain[google-genai]>=0.3.25",
    "langgraph>=0.4.8",
    "python-decouple>=3.8",
]

[tool.uv.workspace]
members = ["app_01/projects/app_01"]
